+++ final_pjt_front/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <link rel="icon" href="/favicon.ico">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>eggmoneyna</title>
+    <title>🐣 Eggmoneyna</title>
 <!-- Bootstrap CSS -->
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css"
@@ -18,14 +18,19 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Jua&family=Nanum+Gothic&display=swap" rel="stylesheet">
-
     
+    <style>
+      @font-face {
+          font-family: 'EnvironmentR';
+          src: url('https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_2206-02@1.0/EnvironmentR.woff2') format('woff2');
+          font-weight: normal;
+          font-style: normal;
+      }
+    </style>
   </head>
 
   <body>
     <div id="app"></div>
-
-    <!-- Vue 앱을 시작하는 main.js 파일을 불러옴 -->
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
