--- final_pjt_front/src/views/SearchView.vue
@@ -1,13 +1,0 @@
-<template>
-  <div>
-
-  </div>
-</template>
-
-<script setup>
-
-</script>
-
-<style lang="scss" scoped>
-
-</style>
